#update total amount row missed

See merge request efms/eFMS-WebApp!2773

+++ WebAPI/eFMS.API.SystemWeb/eFMS.API.ReportData/eFMS.API.ReportData/FormatExcel/AccountingHelper.cs
@@ -656,7 +656,7 @@ namespace eFMS.API.ReportData.FormatExcel
                 for (int i = 0; i < customerPayment.Count; i++)
                 {
                     var item = customerPayment[i];
-                    if (i == 1)
+                    if (i == 1 && isExistDetail)
                     {
                         excel.SetDataTable();
                         excel.DeleteRow(7);
